LibCore+LibIMAP: Move Promise to LibCore
This makes Promise available without having to link LibIMAP.
